# server/docs/swagger_output_dev_new.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1/student/auth/register": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "regNo": {
                  "example": "any"
                },
                "block": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "roomNo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "block": {
                      "type": "string",
                      "example": "A"
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "regNo": {
                      "type": "string",
                      "example": "19BCE0000"
                    },
                    "id": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/student/auth/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "regNo": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student successfully logged in.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Logged in successfully"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "John Wick"
                    },
                    "regNo": {
                      "type": "string",
                      "example": "21BPS1390"
                    },
                    "block": {
                      "type": "string",
                      "example": "A"
                    },
                    "roomNo": {
                      "type": "string",
                      "example": "A-102"
                    },
                    "messType": {
                      "type": "string",
                      "example": "Special"
                    },
                    "toBeChangedTo": {
                      "type": "string",
                      "example": ""
                    },
                    "isForChange": {
                      "type": "boolean",
                      "example": false
                    },
                    "roomId": {
                      "type": "number",
                      "example": 3
                    }
                  }
                },
                "token": {
                  "type": "string",
                  "example": ""
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/v1/warden/auth/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "block": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Warden A"
                    },
                    "block": {
                      "type": "string",
                      "example": "A"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "token": ""
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/v1/faculty/auth/register": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "empId": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "isHOD": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Faculty successfully registered.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Faculty created successfully"
                },
                "faculty": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 5
                    },
                    "empId": {
                      "type": "string",
                      "example": "12347"
                    },
                    "name": {
                      "type": "string",
                      "example": "Faculty-5"
                    },
                    "isHOD": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/faculty/auth/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "empId": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Faculty successfully registered.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Logged in successfully"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 5
                    },
                    "empId": {
                      "type": "string",
                      "example": "12347"
                    },
                    "name": {
                      "type": "string",
                      "example": "Faculty-5"
                    },
                    "isHOD": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                },
                "token": {
                  "type": "string",
                  "example": ""
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/v1/faculty/course/get-all-courses": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Faculty successfully registered.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 2
                      },
                      "courseName": {
                        "type": "string",
                        "example": "DAA"
                      },
                      "courseCode": {
                        "type": "string",
                        "example": "BCSE204L"
                      },
                      "courseType": {
                        "type": "string",
                        "example": "Theory"
                      },
                      "courseCredits": {
                        "type": "number",
                        "example": 4
                      },
                      "courseTeacherId": {
                        "type": "number",
                        "example": 4
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/faculty/course/get-all-students-from-course": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Faculty successfully registered.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "John Wick"
                      },
                      "regNo": {
                        "type": "string",
                        "example": "21BPS1390"
                      },
                      "block": {
                        "type": "string",
                        "example": "A"
                      },
                      "password": {
                        "type": "string",
                        "example": "$2b$10$hifMoCiUGNhY8Imz3chcIebaa5SuRBQPI/VgFOjYnBzy01ejwW4Hy"
                      },
                      "roomNo": {
                        "type": "string",
                        "example": "A-102"
                      },
                      "messType": {
                        "type": "string",
                        "example": "Special"
                      },
                      "toBeChangedTo": {
                        "type": "string",
                        "example": ""
                      },
                      "isForChange": {
                        "type": "boolean",
                        "example": false
                      },
                      "roomId": {
                        "type": "number",
                        "example": 3
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}